Run with multiple profiles

+++ packages/runner/src/index.mjs
@@ -6,12 +6,11 @@ if (arg.option('verbose')) {
   $.verbose = true;
 }
 
-async function _profile(lifecycle) {
-  const {profile} = lifecycle;
+async function _profile(profile) {
   const resp = await fetch(`https://raw.githubusercontent.com/msgport/autoconf/main/ormpipe/runner-${profile}.yml`);
   if (resp.status !== 200) {
-    console.log(chalk.red(`can not read profile ${profile}, please add --profile and there is allow profiles https://github.com/msgport/autoconf/tree/main/ormpipe`));
-    process.exit(1);
+    console.log(chalk.yellow(`can not read profile ${profile}, please add --profile and there is allow profiles https://github.com/msgport/autoconf/tree/main/ormpipe`));
+    return null;
   }
   const body = await resp.text();
   const hash = sha1(body)
@@ -26,39 +25,62 @@ function _extractEnvs() {
   return Object.fromEntries(Object.entries($.env).filter(([k, v]) => k.startsWith('ORMPIPE_')));
 }
 
-async function _start(options = {name, profile}, lifecycle) {
-  const {name, profile} = options;
+async function _start(lifecycle) {
   console.log('===== start');
-  const envs = _extractEnvs();
+  const _definedEnvs = _extractEnvs();
 
-  const ormpipeImageInfo = profile.ormpipe;
-  const {pairs} = ormpipeImageInfo;
+  for (const profileName of lifecycle.profiles) {
 
-  for (const feature of lifecycle.features) {
-    const containerName = `ormpipe-${feature}-${name}`;
+    const profiledSingerEnvName = `ORMPIPE_SIGNER_${profileName.toUpperCase()}`;
+    const envs = {
+      ..._definedEnvs,
+      ORMPIPE_SIGNER: _definedEnvs[profiledSingerEnvName] || _definedEnvs['ORMPIPE_SIGNER'],
+    };
+    delete envs[profiledSingerEnvName];
 
-    const runContainersOutput = await $`docker ps -a --format '{{.Names}}'`.quiet();
-    const rcs = runContainersOutput.stdout.split('\n').filter(item => item);
-    if (rcs.filter(rc => rc === containerName).length) {
-      await $`docker stop ${containerName}`;
-      await $`docker rm ${containerName}`;
+    const profile = await _profile(profileName);
+    if (!profile) {
+      continue;
     }
 
-    await $`docker pull ${ormpipeImageInfo.image}`;
+    if (lifecycle.profileHash[profileName] === profile.hash) {
+      console.log('profile hash not changed');
+      continue
+    }
+    lifecycle.profileHash[profileName] = profile.hash;
+    console.log(`profile changed restart ormpipe program: [${profileName}]`);
 
-    const flags = [
-      '-dit',
-      '--restart=always',
-      `--name=${containerName}`,
-      ...Object.entries(envs).map(([k, v]) => `--env=${k}=${v}`),
-      ormpipeImageInfo.image,
-      feature,
-      ...pairs.map(item => `--enable-pair=${item}`),
-    ]
+    const ormpipePayloadInfo = profile.payload.ormpipe;
+    const {pairs} = ormpipePayloadInfo;
 
-    await $`docker run ${flags}`.quiet();
-    console.log(`-> ormpipe ${chalk.green(containerName)} started with ${chalk.yellow(pairs)}`);
+    for (const feature of lifecycle.features) {
+      const containerName = `ormpipe-${feature}-${profileName}`;
+
+      const runContainersOutput = await $`docker ps -a --format '{{.Names}}'`.quiet();
+      const rcs = runContainersOutput.stdout.split('\n').filter(item => item);
+      if (rcs.filter(rc => rc === containerName).length) {
+        await $`docker stop ${containerName}`;
+        await $`docker rm ${containerName}`;
+      }
+
+      await $`docker pull ${ormpipePayloadInfo.image}`;
+
+      const flags = [
+        '-dit',
+        '--restart=always',
+        `--name=${containerName}`,
+        ...Object.entries(envs).map(([k, v]) => `--env=${k}=${v}`),
+        ormpipePayloadInfo.image,
+        feature,
+        ...pairs.map(item => `--enable-pair=${item}`),
+      ]
+
+      await $`docker run ${flags}`.quiet();
+      console.log(`-> ormpipe ${chalk.green(containerName)} started with ${chalk.yellow(pairs)}`);
+    }
+
   }
+
 }
 
 async function _clean(lifecycle) {
@@ -86,25 +108,18 @@ async function _clean(lifecycle) {
 
 async function main() {
   const lifecycle = {
-    profile: arg.option('profile', 'p'),
-    features: arg.options('features', 'f'),
+    profiles: arg.options('profile', 'p'),
+    features: arg.options('feature', 'f'),
+    profileHash: {},
   };
   if (!lifecycle.features.length) {
     console.log(chalk.yellow('[warn]: no features provided, please use --features or -f to set it'))
   }
 
   while (true) {
-    const {name, hash, payload} = await _profile(lifecycle);
-    if (lifecycle.profileHash === hash) {
-      console.log('profile hash not changed');
-      await setTimeout(1000 * 60 * 2);
-      continue
-    }
-    lifecycle.profileHash = hash;
-    console.log(`profile changed restart ormpipe program: [${name}]`);
-    await _start({name, profile: payload}, lifecycle);
+    await _start(lifecycle);
     await _clean(lifecycle);
-    await setTimeout(1000);
+    await setTimeout(1000 * 60 * 2);
   }
 }
